fix(client): date type conversion
