+++ packages/libs/client/src/index.ts
@@ -1,5 +1,3 @@
-import * as PactValues from './utils/pact-values';
-
 export * from './signing';
 export * from './signing-api/v1/quicksign';
 export * from './signing-api/v1/sign';
@@ -10,7 +8,6 @@ export * from './client';
 export type { IPact, IPactModules } from './pact';
 
 export { Pact } from './pact';
-export { PactValues };
 
 export type * from './interfaces/IPactCommand';
