+++ packages/libs/client/src/utils/pact-helpers.ts
@@ -1,8 +1,4 @@
-/**
- *
- * @alpha
- */
-export type ReadKeyset = <TKey extends string>(
+type ReadKeyset = <TKey extends string>(
   key: TKey,
 ) => () => `(read-keyset "${TKey}")`;
