+++ packages/libs/client/src/utils/parseAsPactValue.ts
@@ -20,10 +20,7 @@ export function parseAsPactValue(
       if ('int' in arg) {
         return new PactNumber(arg.int).toInteger();
       }
-      if (isDate(arg)) {
-        const isoTime = `${arg.toISOString().split('.')[0]}Z`;
-        return `(time "${isoTime}")`;
-      }
+      if (isDate(arg)) return `(time "${arg.toISOString()}")`;
 
       return arg;
     }
